requirements.txt
streamlit
langchain
langchain_chroma
langchain_google_genai
langchain_community
pypdf
python-dotenv
chromadb
sentence-transformers